% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{Get_zeta}
\alias{Get_zeta}
\title{Calculate Covariances of Z-scores between Traits from Overlapping Samples}
\usage{
Get_zeta(common.sumstats, trait.list)
}
\arguments{
\item{common.sumstats}{a numeric list of score summary statistics for all independent common variants that are not associated with the traits}

\item{trait.list}{a vector containing the traits}
}
\value{
A \eqn{K \times K} matrix \eqn{\zeta}, where \eqn{K} is the number of traits.
}
\description{
This function allows you to estimate the matrix \eqn{\zeta} to adjust for the potential sample overlap in the data set. Here we applied LD pruning (\eqn{r^2 < 0.1} in 500kb region) on 1000 Genome genotype dataset (hg19) as a list of reference independent SNPs. The SNP ID is chr:pos.
}
\examples{
\donttest{data(zeta.example)
attach(zeta.example)
# Downloading independent common SNPs from 1000Genome data set.
githubURL <- "https://github.com/lan/MTAR/blob/master/indp_snps.1KG.rda?raw=true"
utils::download.file(githubURL,"1kgfile")
load("1kgfile")
zeta1 <- Get_zeta(dat = dat, indp.common.geno.dat = indp.common.geno.dat,
trait.list = paste0("Trait", 1:K),
cov.list = c("cov1", "cov2"), env = "treatment")
zeta1
detach(zeta.example)
}
}
\author{
Lan Luo
}
