% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGENTA.R, R/import_pkgs.R
\docType{package}
\name{MAGENTA}
\alias{MAGENTA}
\alias{MAGENTA-package}
\title{Calculate MAGENTA omnibus joint or GEI test p value using eq3/eq4}
\usage{
MAGENTA(beta.sumstats.obj, MAF, KA, drugStruct = "AR1", trait.list,
  drug.list, output = "omnibus", MAF.thres = NULL, R.C, cct = TRUE,
  MAC10.bygrp = NULL, test, ref = NULL, diffweight = FALSE, way,
  threshold = 0.05, weight.commonSNP = NULL, rho.SNP = c(0, 0.5, 1),
  rho.trait = c(0.5, 5, 1), rho.drug = c(0, 0.5, 1),
  weight.SNP = c(1, 25))
}
\arguments{
\item{beta.sumstats.obj}{a numeric list of summary statistics generated by Get_beta_cov_UV() or Get_beta_cov_datat() functions in the MAGENTA, depending on the input data}

\item{MAF}{a numeric vector containing the minor allele frequency of the \eqn{M} SNPs in the gene/SNP-set/pathway}

\item{KA}{a \eqn{K \times K} matrix of genetic correlation of the \eqn{K} traits}

\item{drugStruct}{default is "AR1", which is auto-regressive structure with order of 1}

\item{trait.list}{a vector containing the traits}

\item{drug.list}{a vector containing the drugs or environmental groups}

\item{output}{can be set to "omnibus", "MTMV", "STMV", "MTSV", "STSV" or "everything".}

\item{MAF.thres}{A threshold to select SNPs with MAF > thres to be included in the single-variant analysis}

\item{R.C}{a \eqn{M \times M} matrix containing LD correlation among SNPs}

\item{cct}{whether to use Cauchy Combination test to combine the resutls, default is TRUE}

\item{MAC10.bygrp}{a vector containing the SNPs that users want to include in the single-variant analysis}

\item{test}{if set to "joint", the function will conduct joint test; if set to "GEI", the function will return GEI p value}

\item{ref}{the reference environmental group. If empty, it will be set as D (the largest group ID)}

\item{diffweight}{whether use different weight for common and rare variants. Default is false.}

\item{way}{should be set to "eq3" or "eq4"}

\item{threshold}{a threshold to define common or rare variants. Default is 0.05.}

\item{weight.commonSNP}{if diffweight is True, the users should specify the weight function for common SNPs, like dbeta(MAF, 1, 1)}

\item{rho.SNP}{\eqn{rho_S} denotes the tuning parameters for between-SNP signal structure. Default is (0, 0.5,1)}

\item{rho.trait}{\eqn{rho_T} denotes the tuning parameters for between-trait signal structure. Default is (0.5,1, 5)}

\item{rho.drug}{\eqn{rho_E} denotes the tuning parameters for between-group signal structure. Default is (0, 0.5,1)}

\item{weight.SNP}{The weight function for each SNP. Default is dbeta(MAF, 1, 25)}
}
\value{
The function will return the corresponding MAGENTA omnibus p value, or multi-/single-trait multi-/single-variant p values.
}
\description{
This function allows you to calculate MAGENTA joint or GEI test p value using eq3 (start from individual-level data) or eq4 (start from summary statistics)
}
\examples{
\donttest{
#### start from individual-level data
#' data(rawdata)
names(rawdata)
attach(rawdata)
K <- 3
KA <- matrix(c(1, -0.03, 0.3, -0.03, 1, -0.535, 0.3, -0.535, 1), byrow = T, nrow = 3)
rownames(KA) <- colnames(KA) <- paste0("Trait", 1:K)
trait.list = paste0("Trait", 1:K)
drug.list = paste0("Drug", 0:1)
D <- 2
MAF <- colMeans(geno.dat[, -1])/2
R.C <- cor(geno.dat[, -1])
R.C[is.na(R.C)] <- 0
## start from the individual-level data ##
beta.sumstats.obj <- Get_beta_cov_data(geno.dat = geno.dat, dat = dat,
cov.list = c("cov1", "cov2"),
env = "treatment",
D = "ID", type = "continuous",
trait.list = paste0("Trait", 1:K),
SNP.list = colnames(geno.dat)[-1])
names(beta.sumstats.obj)
MAGENTA(beta.sumstats.obj = beta.sumstats.obj,
test = "joint", way = "eq3",MAF = MAF,
 R.C = R.C, KA = KA, output = "everything",
 trait.list = trait.list,
  drug.list = drug.list)
detach(rawdata)

#### start from score summary statistics
data(sumstats.dat)
names(sumstats.dat)
attach(sumstats.dat)
str(common.sumstats)
zeta.ret <- Get_zeta(common.sumstats = common.sumstats, trait.list = paste("Trait", 1:K))
str(sumstats)
#prepare the summary statistics for joint test using trait-specific score summary statistics
beta.sumstats.obj <- Get_beta_cov_UV(sumstats = sumstats, MAF = MAF, R.C = R.C,
                                     zeta.ret = zeta.ret, KA = KA,
                                     test = "joint",
                                     trait.list = paste0("Trait", 1:K))
names(beta.sumstats.obj)
MAGENTA(beta.sumstats.obj = beta.sumstats.obj,
        test = "joint", way = "eq4",MAF = MAF,MAC10.bygrp = joint.MAC10,
        R.C = R.C, KA = KA, output = "everything",
        trait.list = trait.list,
        drug.list = drug.list)
detach(sumstats.dat)
}
}
\author{
Lan Luo

Lan Luo
}
