% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Example1: individual-level data}
\format{An object of class \code{list} of length 2.}
\usage{
data(rawdata)
}
\description{
An example of calculating summary statistics in MAGENTA given the individual-level data set.
}
\keyword{datasets}
