% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{Get_beta_cov_UV}
\alias{Get_beta_cov_UV}
\title{Calculate the summary statistics required by MAGENTA given the score summary statistics}
\usage{
Get_beta_cov_UV(sumstats, MAF, zeta.ret, R.C, KA, MAF.thres = 0.000588,
  trait.list, ref = NULL, main.thres = 0.005, test)
}
\arguments{
\item{sumstats}{a list of score summary statistics (U, V) in each environmental group for each trait}

\item{MAF}{a numeric vector containing the minor allele frequency of the \eqn{M} SNPs in the gene/SNP-set/pathway}

\item{R.C}{a \eqn{M \times M} matrix containing LD correlation among SNPs}

\item{KA}{a \eqn{K \times K} matrix of genetic correlation of the \eqn{K} traits}

\item{MAF.thres}{A threshold to select SNPs with MAF > thres to be included in the single-variant analysis}

\item{trait.list}{a vector containing the traits}

\item{ref}{the reference environmental group. If empty, it will be set as D (the largest group ID)}

\item{main.thres}{a threshold to prompt warning if the genetic main effect is too strong. Default is 5e-3}

\item{test}{if set to "joint", the function will conduct joint test; if set to "GEI", the function will return GEI p value}
}
\value{
summary statistics (beta, beta.cov) required by MAGENTA
}
\description{
This function allows you to calculate the summary statistics (beta, beta.cov) given the score summary statistics.
}
\examples{
\donttest{
data(sumstats.dat)
names(sumstats.dat)
attach(sumstats.dat)
str(common.sumstats)
zeta.ret <- Get_zeta(common.sumstats = common.sumstats, trait.list = paste("Trait", 1:K))
# zeta.ret
str(sumstats) ## variant-level score summary statistics U and V as input data, generated by cal.sum, zero means that SNP is a non-polymorphic SNP for trait k and environmental group d
beta.sumstats.obj <- Get_beta_cov_UV(sumstats = sumstats, MAF = MAF, R.C = R.C,
                                     zeta.ret = zeta.ret, KA = KA,
                                     test = "joint",
                                   trait.list = paste0("Trait", 1:K))
names(beta.sumstats.obj)
detach(sumstats.dat)
}
}
\author{
Lan Luo
}
