% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstats.dat}
\alias{sumstats.dat}
\title{Example2: the score summary statistics and their variance}
\format{An object of class \code{list} of length 5.}
\usage{
data(sumstats.dat)
}
\description{
An example of calculating summary statistics in MAGENTA given the score summary statistics U and its variance V.
}
\keyword{datasets}
